Testing new Trial implementation final phase

+++ code/pheno_count/app/src/main/java/com/cmput301w21t36/phenocount/NonNegativeCount.java
@@ -16,7 +16,7 @@ public class NonNegativeCount extends Trial implements Serializable {
     }
 
 
-    public float getValue() {
+    public int getValue() {
         return value;
     }
